[rs] Split `parser::movie` into `streaming::movie` and `complete::movie`
